🧪 Update `wandbfsspec` imports
To use `wandbfsspec.spec` instead of `wandbfsspec.core`, as well as removing `rm_file` unit tests for `WandbArtifactStore`
